Add Assess Mgmt API.

+++ yy-common/yy-common-utils/src/main/java/com/yy/cloud/common/data/PageInfo.java
@@ -47,5 +47,5 @@ public class PageInfo implements Serializable {
 	/**
 	 * Optional. totalRecords: total records of the entities.
 	 */
-	private Integer totalRecords;
+	private Long totalRecords;
 }
